Refactor gallery viewer to its own component
